Revert "4 feature 3 fulfilling site drug order requests"
