+++ src/main/java/drug_request_entity/DrugRequest.java
@@ -1,10 +1,7 @@
-package site_drug_request.drug_request_entity;
+package drug_request_entity;
 
 public interface DrugRequest {
-    boolean drugNameIsEmpty();
-    boolean drugBottleIsEmpty();
     boolean drugBottleIsValid();
-    boolean drugBottleIsNumeric();
     boolean drugNameIsValid();
     String getDrugName();
     String getDrugBottle();
