+++ src/main/java/drug_request_entity/DrugRequestGenerator.java
@@ -1,4 +1,4 @@
-package site_drug_request.drug_request_entity;
+package drug_request_entity;
 
 //Reason why create method exists in DrugRequestGenerator is due to the stable abstractions principal (packages that
 // are maximally stable should be maximally abstract).
