+++ src/main/java/drug_request_frameworks_drivers/DrugRequestFailed.java
@@ -1,0 +1,7 @@
+package drug_request_frameworks_drivers;
+
+public class DrugRequestFailed extends RuntimeException{
+    public DrugRequestFailed(String error){
+        super(error);
+    }
+}
