+++ src/main/java/drug_request_frameworks_drivers/PanelGenerator.java
@@ -1,4 +1,4 @@
-package site_drug_request.drug_request_frameworks_drivers;
+package drug_request_frameworks_drivers;
 
 import javax.swing.*;
