+++ src/main/java/drug_request_interface_adapters/DrugRequestController.java
@@ -1,8 +1,8 @@
-package site_drug_request.drug_request_interface_adapters;
+package drug_request_interface_adapters;
 
-import site_drug_request.drug_request_use_case.DrugRequestInputBoundary;
-import site_drug_request.drug_request_use_case.DrugRequestInvokeModel;
-import site_drug_request.drug_request_use_case.DrugRequestResponseModel;
+import drug_request_use_case.DrugRequestInputBoundary;
+import drug_request_use_case.DrugRequestInvokeModel;
+import drug_request_use_case.DrugRequestResponseModel;
 
 public class DrugRequestController {
      DrugRequestInputBoundary drugRequestInputBoundary;
@@ -14,6 +14,5 @@ public class DrugRequestController {
     public DrugRequestResponseModel create(String drugName, String drugBottle){
         DrugRequestInvokeModel drugRequestInvokeModel = new DrugRequestInvokeModel(drugName, drugBottle);
         return drugRequestInputBoundary.create(drugRequestInvokeModel);
-//        return null;
     }
 }
