+++ src/main/java/drug_request_use_case/DrugRequestDsGateway.java
@@ -1,0 +1,6 @@
+package drug_request_use_case;
+
+public interface DrugRequestDsGateway {
+    boolean drugNameExists(String identifier);
+    void saveDrugRequest(DrugRequestDsInvokeModel drugRequestDsInvokeModel);
+}
