+++ src/main/java/drug_request_use_case/JpaUserRepository.java
@@ -1,0 +1,8 @@
+package drug_request_use_case;
+
+import drug_request_interface_adapters.DrugRequestDataMapper;
+
+
+//public interface JpaUserRepository extends JpaRepository<DrugRequestDataMapper>{
+//
+//}
