--- src/main/java/site_drug_request/drug_request_frameworks_drivers/DrugRequestFailed.java
@@ -1,11 +1,0 @@
-package site_drug_request.drug_request_frameworks_drivers;
-
-public class
-
-
-
-DrugRequestFailed extends RuntimeException{
-    public DrugRequestFailed(String error){
-        super(error);
-    }
-}
