--- src/main/java/site_drug_request/drug_request_use_case/DrugRequestDsGateway.java
@@ -1,6 +1,0 @@
-package site_drug_request.drug_request_use_case;
-
-public interface DrugRequestDsGateway {
-    boolean drugNameExists(String identifier);
-    void generateDrugRequest(DrugRequestDsInvokeModel drugRequestDsInvokeModel);
-}
