--- src/main/java/site_drug_request/drug_request_use_case/JpaUserRepository.java
@@ -1,6 +1,0 @@
-package site_drug_request.drug_request_use_case;
-
-
-//public interface JpaUserRepository extends JpaRepository<DrugRequestDataMapper>{
-//
-//}
